requirements.txt
flask
flask-bootstrap4
flask-wtf